separated Magbead and Pipette to different files
